Refactor organisation to prevent cyclic imports
