+++ reiz/_marker/mitm.py
@@ -17,7 +17,7 @@ import queue
 import time
 import socket
 import json
-from reiz.marker.client import available
+from reiz._marker.client import available
 # %%
