--- reiz/marker/__init__.py
@@ -1,2 +1,0 @@
-from reiz.marker.client import push, push_json, available
-from reiz.marker.safeguard import start, stop
